+++ packages/admin/cms-admin/src/pages/pageTree/PageTreeContext.ts
@@ -1,3 +1,4 @@
+import { DocumentNode } from "graphql";
 import * as React from "react";
 
 import { DocumentInterface, DocumentType } from "../../documents/types";
@@ -10,6 +11,7 @@ export interface PageTreeContext {
     allCategories: AllCategories;
     documentTypes: Record<DocumentType, DocumentInterface>;
     tree: TreeMap<GQLPageTreePageFragment>;
+    query: DocumentNode;
 }
 
 export const PageTreeContext = React.createContext<PageTreeContext | undefined>(undefined);